Update changelog
This commit revives project's changelog. The idea behind it is
following: build tests as a single jar file and provide it to users
(like Tarantool's CI and personal use). In such case it would simplify
using tests and will not require setup of build enviroment with
Leiningen and Clojure. To distinguish different versions we need
versioning and changelog.

Related to #97